genirq/msi-doorbell: creation
This new API aims at managing MSI doorbells likely to be iommu
mapped. The API is bound to be used at least by irqchip drivers,
the MSI layer and the VFIO sub-system.

Irqchips likely to be downstream to IOMMUs (that do not bypass MSI
transactions) need to register their MSI doorbells so that the MSI
layer can look them up and iommu map then when needed.

We currently introduce the irq_chip_msi_doorbell_info struct that
contains all the info characterizing a doorbell plus 3 functions that
allows to (un)register, and lookup MSI doorbells stored in a global
list.

Signed-off-by: Eric Auger <eric.auger@redhat.com>