genirq/msi: Introduce msi_desc flags
This new flags member is meant to store additional information about
the msi descriptor, starting with allocation status information.

MSI_DESC_FLAG_ALLOCATED bit tells the associated base IRQ is allocated.
This information is currently used at deallocation time. We also
introduce MSI_DESC_FLAG_FUNCTIONAL telling the MSIs are functional.

For the time being ALLOCATED and FUNCTIONAL are set at the same time
but this is going to change in subsequent patch. Indeed in some situations
some additional tasks need to be carried out for the MSI to be functional.
For instance the MSI doorbell may need to be mapped in an IOMMU.

FUNCTIONAL value already gets used when enumerating the usable MSIs in
msix_capability_init.

Signed-off-by: Eric Auger <eric.auger@redhat.com>

---

v12: new