VFIO: pci: Introduce direct EOI INTx interrupt handler
We add two new fields in vfio_pci_irq_ctx struct: deoi and handler.
If deoi is set, this means the physical IRQ attached to the virtual
IRQ is directly deactivated by the guest and the VFIO driver does
not need to disable the physical IRQ and mask it at VFIO level.

The handler pointer is set accordingly and a wrapper handler is
introduced that calls the chosen handler function.

Signed-off-by: Eric Auger <eric.auger@redhat.com>

---