KVM: arm/arm64: Record RDIST Last bit at registration
To prepare for multiple RDIST regions, let's record the RDIST
Last bit field when registering the IO device.

As a reminder the Last bit indicates whether the redistributor
is the highest one in a series of contiguous redistributor pages.

Since at the moment KVM only supports a single redist region,
the RDIST tagged with the last bit set to true corresponds to the
highest vCPU one.

Signed-off-by: Eric Auger <eric.auger@redhat.com>